一応予備動作できたけどコードがよろしくないため改良の余地あり

+++ Game/Objects/Stage/Components/TimerComponent/Component_Timer.h
@@ -78,7 +78,7 @@ public:
 	/// <param name="_time">
 	/// カウントする時間(秒)
 	/// </param>
-	void SetTime(int _time);
+	void SetTime(float _time);
 
 	/// <summary>
 	/// タイマーをリセット
